# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Bike Auction Platform API
  description: |
    RESTful API for the Bike Auction Platform.

    ## Authentication
    - **Register/Login** → Receive a short-lived JWT Access Token (15 min) in the response body
      and a long-lived Refresh Token (7 days) as an `httpOnly`, `SameSite=Strict`, `Secure` cookie.
    - **Authenticated requests** → Pass the Access Token in the `Authorization: Bearer <token>` header.
    - **Token refresh** → `POST /api/auth/refresh` reads the cookie and returns a new Access Token.
    - **Real-time bidding** → Uses WebSocket (Socket.io) with token passed in the connection handshake.
      See the architecture document for WebSocket event contracts.

    ## Monetary Values
    All monetary fields are represented in **cents** (integer) to avoid floating-point precision issues.
    Example: `$1,250.00` → `125000`.
  version: "1.0.0"
  contact:
    name: Bike Auction Engineering
    email: eng@bikeauction.dev
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.bikeauction.dev
    description: Production

# ============================================================================
# Security Schemes
# ============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived access token (15 minutes). Obtain via POST /api/auth/login.

  # ==========================================================================
  # Reusable Schemas
  # ==========================================================================
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: "Email is required"
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string

    # --- Auth ---
...
    MotorcycleCreate:
      type: object
      required: [title, make, model, year, mileageKm, condition]
      properties:
        title:
          type: string
          maxLength: 300
          example: "2023 Ducati Panigale V4 S — Pristine Condition"
        make:
          type: string
          maxLength: 100
          example: Ducati
        model:
          type: string
          maxLength: 100
          example: Panigale V4 S
        year:
          type: integer
          minimum: 1900
          maximum: 2100
          example: 2023
        vin:
          type: string
          maxLength: 17
          example: ZDM14BKW1PB000123
        mileageKm:
          type: integer
          minimum: 0
          example: 3200
        engineCc:
          type: integer
          example: 1103
        condition:
          type: string
          enum: [excellent, good, fair, poor, salvage]
          example: excellent
        description:
          type: string
          example: "One owner, full service history, never dropped."
        photoUrls:
          type: array
          items:
            type: string
            format: uri
          example:
            - https://cdn.bikeauction.dev/photos/ducati-front.jpg
            - https://cdn.bikeauction.dev/photos/ducati-side.jpg
        thumbnailUrl:
          type: string
          format: uri
    MotorcycleUpdate:
      type: object
      properties:
        title:
          type: string
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        vin:
          type: string
        mileageKm:
          type: integer
        engineCc:
          type: integer
        condition:
          type: string
          enum: [excellent, good, fair, poor, salvage]
        description:
          type: string
        photoUrls:
          type: array
          items:
            type: string
            format: uri
        thumbnailUrl:
          type: string
          format: uri
    Motorcycle:
      allOf:
        - $ref: "#/components/schemas/MotorcycleCreate"
        - type: object
          properties:
            id:
              type: string
              format: uuid
            sellerId:
              type: string
              format: uuid
            listingStatus:
              type: string
              enum: [draft, pending_approval, approved, rejected]
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time

    # --- Auction ---
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalItems:
          type: integer
          example: 142
        totalPages:
          type: integer
          example: 8

  # ==========================================================================
  # Reusable Parameters
  # ==========================================================================
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    QueryPage:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    QueryLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20

# ============================================================================
# Paths
# ============================================================================
paths:
...
  /api/motorcycles:
    get:
      tags: [Motorcycles]
      summary: List approved motorcycles (public browse)
      operationId: listMotorcycles
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - name: make
          in: query
          schema:
            type: string
          description: "Filter by manufacturer (e.g., Ducati, Harley-Davidson)"
        - name: yearMin
          in: query
          schema:
            type: integer
        - name: yearMax
          in: query
          schema:
            type: integer
        - name: condition
          in: query
          schema:
            type: string
            enum: [excellent, good, fair, poor, salvage]
        - name: sort
          in: query
          schema:
            type: string
            enum: [newest, oldest, year_asc, year_desc]
            default: newest
      responses:
        "200":
          description: Paginated list of approved motorcycles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Motorcycle"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"

    post:
      tags: [Motorcycles]
      summary: Create a motorcycle listing (seller only)
      operationId: createMotorcycle
      security:
        - BearerAuth: []
      description: |
        Requires `role: seller` and `isSellerVerified: true`.
        Listing is created in `draft` status.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotorcycleCreate"
      responses:
        "201":
          description: Motorcycle listing created in draft status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not a verified seller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/motorcycles/{id}:
    get:
      tags: [Motorcycles]
      summary: Get motorcycle details
      operationId: getMotorcycle
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Motorcycle details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
        "404":
          description: Motorcycle not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      tags: [Motorcycles]
      summary: Update motorcycle listing (owner only)
      operationId: updateMotorcycle
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      description: |
        Only the listing owner can update. Cannot update if listing
        has an active auction.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MotorcycleUpdate"
      responses:
        "200":
          description: Updated motorcycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
        "403":
          description: Not the listing owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Motorcycle not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot update — motorcycle has an active auction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Motorcycles]
      summary: Delete motorcycle listing (owner only)
      operationId: deleteMotorcycle
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      responses:
        "200":
          description: Motorcycle deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Motorcycle listing deleted
        "403":
          description: Not the listing owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Cannot delete — motorcycle has an active auction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/motorcycles/{id}/submit:
    post:
      tags: [Motorcycles]
      summary: Submit listing for approval (owner only)
      operationId: submitMotorcycleForApproval
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PathId"
      description: Transitions listing from `draft` → `pending_approval`.
      responses:
        "200":
          description: Listing submitted for approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Motorcycle"
        "400":
          description: Listing not in draft status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not the listing owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # --------------------------------------------------------------------------
  # Auctions
  # --------------------------------------------------------------------------
...
  /api/users/me/motorcycles:
    get:
      tags: [Users]
      summary: Get my motorcycle listings (seller dashboard)
      operationId: getMyMotorcycles
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/QueryPage"
        - $ref: "#/components/parameters/QueryLimit"
        - name: listingStatus
          in: query
          schema:
            type: string
            enum: [draft, pending_approval, approved, rejected]
      responses:
        "200":
          description: Paginated list of the seller's motorcycle listings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Motorcycle"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
...